uint/int to usize/isize, {:?}

+++ src/util.rs
@@ -1,4 +1,4 @@
-pub fn u64_to_le_bytes(n: u64, nbytes: uint) -> Vec<u8> {
+pub fn u64_to_le_bytes(n: u64, nbytes: usize) -> Vec<u8> {
     let mut bytes = Vec::with_capacity(nbytes);
     for i in range(0, nbytes) {
         bytes.push((n >> (i * 8)) as u8);
@@ -16,7 +16,7 @@ pub fn bytes_to_le_u64(bytes: &[u8]) -> u64 {
     n
 }
 
-pub fn u64_to_be_bytes(n: u64, nbytes: uint) -> Vec<u8> {
+pub fn u64_to_be_bytes(n: u64, nbytes: usize) -> Vec<u8> {
     let mut bytes = Vec::with_capacity(nbytes);
     for i in range(1, nbytes + 1) {
         bytes.push((n >> ((nbytes - i) * 8)) as u8);
